- add view binding - add loading layout
