+++ app/src/main/java/com/kedra/filedownloader/main/view/adapter/MainAdapter.kt
@@ -1,0 +1,24 @@
+package com.kedra.filedownloader.main.view.adapter
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import com.kedra.filedownloader.R
+
+class MainAdapter : RecyclerView.Adapter<MainAdapter.ItemHolder>() {
+
+    inner class ItemHolder(itemView: View) : RecyclerView.ViewHolder(itemView)
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ItemHolder {
+        val view =
+            LayoutInflater.from(parent.context).inflate(R.layout.item_list_layout, parent, false)
+        return ItemHolder(view)
+    }
+
+    override fun onBindViewHolder(holder: ItemHolder, position: Int) {
+
+    }
+
+    override fun getItemCount(): Int = 12
+}
