+++ app/src/main/java/com/kedra/filedownloader/main/view/ui/MainFragment.kt
@@ -1,0 +1,43 @@
+package com.kedra.filedownloader.main.view.ui
+
+import androidx.lifecycle.ViewModelProvider
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.kedra.filedownloader.R
+import com.kedra.filedownloader.databinding.MainFragmentBinding
+import com.kedra.filedownloader.main.view.adapter.MainAdapter
+import com.kedra.filedownloader.main.viewModel.MainViewModel
+
+class MainFragment : Fragment() {
+
+    companion object {
+        fun newInstance() = MainFragment()
+    }
+
+    private lateinit var viewModel: MainViewModel
+
+    private lateinit var binding: MainFragmentBinding
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        binding = MainFragmentBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        initUi()
+    }
+
+    private fun initUi(){
+        with(binding){
+            rvList.adapter = MainAdapter()
+        }
+    }
+
+}
