+++ app/src/main/java/com/kedra/filedownloader/main/viewModel/MainViewModel.kt
@@ -1,4 +1,4 @@
-package com.kedra.filedownloader.ui.main
+package com.kedra.filedownloader.main.viewModel
 
 import androidx.lifecycle.ViewModel
