+++ app/src/main/res/layout/item_list_layout.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_margin="16dp"
-    android:elevation="8dp"
+    android:elevation="16dp"
     android:layout_height="wrap_content">
 
     <LinearLayout
